LayoutTests/imported/w3c/web-platform-tests/css/css-view-transitions/navigation/pageswap-long-delay-expected.txt


FAIL View transitions: long delay in pageswap aborts the transition assert_equals: viewTransition must have been skipped. expected null but got object "[object ViewTransition]"
